pyproject.toml
[project]
name = "document-knowledge-pipe"
version = "0.1.0"
description = "Document ingestion pipeline: folder → markdown → knowledge graph"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Hannes Suhr" }]
dependencies = [
    "pymupdf4llm>=0.0.17",
    "openai>=1.40.0",
    "anthropic>=0.40.0",
    "lightrag-hku>=1.0.0",
    "watchdog>=4.0.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.scripts]
docpipe = "docpipe.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docpipe"]

[tool.hatch.build.targets.sdist]
exclude = [".env", "output/", "logs/", "input/"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["lightrag.*", "pymupdf4llm.*", "fitz.*", "yaml.*", "watchdog.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "needs_anthropic_key: requires ANTHROPIC_API_KEY",
    "needs_openai_key: requires OPENAI_API_KEY",
    "needs_api_keys: requires both ANTHROPIC_API_KEY and OPENAI_API_KEY",
    "needs_libreoffice: requires LibreOffice installed",
]
